[sapphire] Rename ServerJob -> SapphireJob
