+++ sapphire/sapphire_worker.py
@@ -2,6 +2,9 @@
 # This Source Code Form is subject to the terms of the Mozilla Public
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
+"""
+Sapphire HTTP server worker
+"""
 
 import logging
 import socket
