+++ sapphire/status_codes.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+SERVED_ALL = 0      # all expected requests for required files have been received
+SERVED_NONE = 1     # no requests for required files have been received
+SERVED_REQUEST = 2  # some requests for required files have been received
+SERVED_TIMEOUT = 3  # timeout occurred
